Added Jakarta tests to GenericModelTypesResourceTest Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
