+++ extension-jaxrs/src/test/java/test/io/smallrye/openapi/runtime/scanner/jakarta/KingCrimson.java
@@ -1,0 +1,47 @@
+package test.io.smallrye.openapi.runtime.scanner.jakarta;
+
+import java.util.Date;
+
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+import jakarta.json.bind.annotation.JsonbDateFormat;
+
+public class KingCrimson extends BaseModel {
+
+    public enum Status {
+        unknown,
+        success,
+        failure
+    }
+
+    @JsonbDateFormat(value = "yyyy-MM-dd'T'HH:mm:ss[.SSS]X")
+    @Schema(implementation = String.class, format = "date-time")
+    Date timestamp;
+    Magma environment;
+    Status status;
+
+    public Date getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(Date timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public Magma getEnvironment() {
+        return environment;
+    }
+
+    public void setEnvironment(Magma environment) {
+        this.environment = environment;
+    }
+
+    public Status getStatus() {
+        return status;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+
+}
