+++ extension-jaxrs/src/test/java/test/io/smallrye/openapi/runtime/scanner/jakarta/POJO.java
@@ -1,0 +1,15 @@
+package test.io.smallrye.openapi.runtime.scanner.jakarta;
+
+public class POJO extends BaseModel {
+
+    String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
