+++ extension-jaxrs/src/test/java/test/io/smallrye/openapi/runtime/scanner/jakarta/Residents.java
@@ -1,0 +1,24 @@
+package test.io.smallrye.openapi.runtime.scanner.jakarta;
+
+public class Residents extends BaseModel {
+
+    String foo;
+    String bar;
+
+    public String getFoo() {
+        return foo;
+    }
+
+    public void setFoo(String foo) {
+        this.foo = foo;
+    }
+
+    public String getBar() {
+        return bar;
+    }
+
+    public void setBar(String bar) {
+        this.bar = bar;
+    }
+
+}
